Address starter flow review feedback
