+++ template/Tuist/Interfaces/SwiftUI/Sources/Presentation/Modules/SignOut/SignOutView.swift
@@ -5,7 +5,7 @@ struct SignOutView: View {
     var onContinue: () -> Void = {}
 
     var body: some View {
-        VStack(spacing: 20) {
+        VStack(spacing: 20.0) {
             Image(systemName: "person.crop.circle.badge.exclamationmark")
                 .font(.system(size: 48))
                 .foregroundColor(.accentColor)
